pyproject.toml
[project]
name = "repl-nix-ytubefen-xi-matupu"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client>=2.153.0",
    "google-generativeai>=0.8.3",
    "isodate>=0.7.2",
    "networkx>=3.4.2",
    "plotly>=5.24.1",
    "streamlit>=1.40.1",
    "youtube-transcript-api>=0.6.2",
]
